+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/bot/settings/limiters/EmptyLimiter.kt
@@ -1,5 +1,0 @@
-package dev.inmo.tgbotapi.bot.settings.limiters
-
-object EmptyLimiter : RequestLimiter {
-    override suspend fun <T> limit(block: suspend () -> T): T = block()
-}